AES-GCM encryption with nonce = HMAC(title)
I've successfully set up the paradigm for AES-GCM, and all of our notes now run through this encryption/decryption schema. I am currently using HMAC of the title as the nonce, but this might be bad practice. Should be trivial to swap out if so. I will also likely make a new commit with less awful comments a bit later.