PKCS7 padding/unpadding
Padding and unpadding has been added before encrypting and after decrypting AES-GCM. This is to ensure that ciphertexts stay the same length. A little shaky on my definition of byte length here? It seems like the preset length of a note is 2048 but PKCS7 can only pad to a block size of up to 2040. Gonna ask in office hours if this creates any vulnerability edge cases that must be handled.