(Possibly naive) checksum implementation
Dump now gives a checksum equal to the SHA256 hash of the serialized data. Init checks equality between this and a new calculated hash of the data, to protect against rollback. Not sure if this is like, correct? But it sure is in the codebase now! Wow. I really am a software developer.