unpadded data length title is now MAC of the hash
hash of the MAC seems to be on shaky ground when we're already using the MAC as the titles of our notes, so the title for these is now the MAC of the hash. MAC being on the top level should ensure pseudorandom output that can't be used to discern what's a message and what's a length value.